Refactor migration code to resolve linter issues and improve code quality
- Updated versions.json with new hash values for migration scripts.
- Addressed linter noise in migration functions, specifically:
  - Q1: Fixed false positives in `create_dd_table` and `enable_dd_table` by adjusting format templates.
  - Q2: Converted `SELECT expr INTO variable` to plain assignments in multiple functions.
  - Q3: Made implicit casts explicit in various functions.
  - Q4: Renamed shadowed loop variables in `evaluate_json_logic` to eliminate warnings.
  - Q5: Removed unused variables and re-linted functions to ensure cleanliness.
- Closed Q1 and Q4; restated Q2, Q3, and Q5 as pgmq-only due to findings in vendored code.
- Documented changes and decisions in ext-solved-items.md and pg_semantius-open-items.md.

+++ apps/_core/migrations/0270_entity_order_column.sql
@@ -77,7 +77,7 @@ BEGIN
     IF v_val IS NULL OR v_val = 0 THEN
         IF TG_TABLE_NAME = 'fields' THEN
             -- Per table_name: a new field lands after that entity's existing fields.
-            SELECT COALESCE(MAX(field_order), 0) + 10
+            SELECT (COALESCE(MAX(field_order), 0) + 10)::BIGINT
             INTO v_next
             FROM fields
             WHERE field_order < 900000
